backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
google-generativeai>=0.8.0
requests>=2.32.0
python-dotenv>=1.0.0
pydantic>=2.8.0
httpx>=0.27.0
beautifulsoup4>=4.12.0